+++ src/main/java/be/baur/sds/NodeType.java
@@ -6,7 +6,6 @@ import be.baur.sda.Node;
 import be.baur.sda.SDA;
 import be.baur.sda.DataNode;
 import be.baur.sds.content.AbstractStringType;
-import be.baur.sds.content.AnyType;
 import be.baur.sds.content.RangedType;
 import be.baur.sds.model.ModelGroup;
 import be.baur.sds.serialization.Attribute;
@@ -25,16 +24,17 @@ import be.baur.sds.serialization.Components;
  * @see ModelGroup
  * @see DataType
  **/
-public class NodeType extends Component {
+public class NodeType extends Type {
 
 	private NodeType globalType = null; // the global type this type refers to, if any
-	private String typeName; // the name of this type
 
 
 	/**
-	 * Creates a node type with the specified name.
+	 * Creates a type with the specified name, which must be a valid node name.
 	 * 
-	 * @param name a valid node name, see {@link SDA#isName}
+	 * @see SDA#isName
+	 * 
+	 * @param name a valid node name
 	 * @throws IllegalArgumentException if the name is invalid
 	 */
 	public NodeType(String name) {
@@ -43,26 +43,31 @@ public class NodeType extends Component {
 	
 
 	/**
-	 * Returns the name of this type.
+	 * Returns the name of this type. A type defines an instance of a data node, so
+	 * this method always returns a valid node name.
 	 * 
 	 * @return a valid node name, not null or empty
 	 */
-	public String getTypeName() {
-		return typeName;
+	@Override
+	public final String getTypeName() {
+		return super.getTypeName();
 	}
 
-
+	
 	/**
-	 * Sets the name of this type. Ultimately, a type represents an instance of a
-	 * data node, so the name of the type is restricted to valid node names.
+	 * Sets the name of this type. A type defines an instance of a data node, so the
+	 * name of this type is restricted to valid node names.
 	 * 
-	 * @param name a valid node name, see {@link SDA#isName}
+	 * @see SDA#isName
+	 * 
+	 * @param name a valid node name, may be null or empty
 	 * @throws IllegalArgumentException if the name is invalid
 	 */
+	@Override
 	public final void setTypeName(String name) {
 		if (! SDA.isName(name)) 
 			throw new IllegalArgumentException("invalid node name (" + name + ")");
-		typeName = name;
+		super.setTypeName(name);
 	}
 
 
@@ -109,11 +114,9 @@ public class NodeType extends Component {
 		
 		final DataNode node = new DataNode(Components.NODE.tag);
 		
-		// Omit the name for an unnamed any type, and for a type
-		// reference with the same name as the referenced type.
-		if (! (( getGlobalType() != null && typeName.equals(getGlobalType()) )
-				|| ( this instanceof AnyType && !((AnyType) this).isNamed() )) ) {
-			node.setValue(typeName);
+		// Omit the name for a reference with the same name as the referenced type
+		if (! ( getGlobalType() != null && getTypeName().equals(getGlobalType()) )) {
+			node.setValue(getTypeName());
 		}
 	
 		// Render the type attribute for a global type reference, or a data type
@@ -122,7 +125,7 @@ public class NodeType extends Component {
 		else if (this instanceof DataType)
 			node.add(new DataNode(Attribute.TYPE.tag, ((DataType) this).getType()));
 		
-		// Render the multiplicity if not default.
+		// Render the multiplicity if not default
 		if (getMultiplicity().min != 1 || getMultiplicity().max != 1) 
 			node.add(new DataNode(Attribute.OCCURS.tag, getMultiplicity().toString()));