+++ src/main/java/be/baur/sds/content/BooleanType.java
@@ -8,7 +8,7 @@ import be.baur.sds.DataType;
  */
 public final class BooleanType extends DataType {
 
-	public static final String TYPE = "boolean";
+	public static final String NAME = "boolean";
 	public static final String TRUE = "true";
 	public static final String FALSE = "false";
 
@@ -37,7 +37,8 @@ public final class BooleanType extends DataType {
 	}
 
 
+	@Override
 	public String getType() {
-		return TYPE;
+		return NAME;
 	}
 }